Support automatically building as if PublicRelease=true
The version.json file can now specify regex patterns that if the ref being built matches will cause PublicRelease to default to true instead of false.

This comes with specialized support for a couple of build server platforms (VSTS and AppVeyor). But it works with pure git as well, both locally on dev boxes and on these and other build servers.

Fix #38