Minimize calls to git when calculating VersionOracle
The VersionOracle constructor was using several helper methods from
GitExtensions to calculate the VersionHeight and Version for the
oracle. Unfortunately, most of the helper methods recalculate the
VersionOptions by querying git and reserializing the version.json
file. This is particularly costly when the git history gets deep.

With this commit, the VersionOracle calculates the commit and working
tree VersionOptions each once and goes about its processing from there.
This should greatly improve performance on deep git history's.

Addresses part of #114

Note: caching VersionOracle across projects in a single MSBuild
invocation still makes a lot of sense.