Remove the `g` prefix in commit IDs almost everywhere
This removes `g` basically everywhere except for semver 1.0 compliant NuGet package versions, since older NuGet clients can't handle -prerelease tags that start with a digit (which some commit IDs do). For these, we still prefix the `g` as the convention to ensure that the prerelease tag starts with a letter.

Fixes #260