Abort prepare-release command when version on current branch is not incremented
Ensure version on current branch is updated by prepare-release command

When the next version for the current branch is set by the user using
the '--nextVersion' command line parameter, the next version might be
equal to the current version.
In that case no commit is created (no changes on the branch), which in
turn causes the current branch to be fast-forwarded to the commit on
the newly-created release branch when the branches are merged.

To prevent this error, abort command if both versions match (there is
already a check in place that prevents the version from being decremented).

The check only applies to the current branch because the version on the
release branch not being incremented is a valid use case.