Always decode the first two bytes of a commit ID into `ushort` as big endian
This fixes a non-deterministic version computation that varied based on endianness of the processor. Most computers use little-endian, which produced a decimal version that when converted back to hex with a typical calculator would have the first two commit ID bytes in swapped positions.
Now with this change, we "fix" the endianness. We choose big endian since it won't swap the first two bytes like little endian did.

But this presents another problem: all the generated versions now have a different value for the 4th integer component, and `nbgv get-commits` will no longer match a commit when it built using the little-endian version of NB.GV. I'll fix this in a subsequent commit so that this CLI tool will match on commits allowing the order to be swapped.

Fixes #637