Include all four version components in the cloud build number if three or more components are set in the version file
If the version in version.json specifies three or four version components, include all four version fields in the cloud build number.
If three components are set in the version file, the git height is encoded as fourth component and should thus be included to make the cloud build number unique.
If four components are set, include the fourth component to ensure the cloud build number is at least as precise as the value in the version file.