Add git-based detection of tags at HEAD to improve PublicRelease detection (#876)
* Add wip tag lookup

* Managed git does not support annotated tags yet
* json output reads: "NBGV_BuildingTags": "System.Collections.Generic.List`1[System.String]",

* Add managed git support for annotated tags

* Fix nullable warnings

* Fix mistaken comment in GitCommitReader

* Fix nullability issues and code style

* Add workaround for improper GitPack.TryGetValue function

* Make HeadTags return null instead of throwing if no HEAD

Throwing was intentional to match HeadCanonicalName behaviour. Throwing fails the tests though. Maybe this is because the tests run in a cloud build environment and HeadCanonicalName isn't used there?

this.BuildingRef = cloudBuild?.BuildingTag ?? cloudBuild?.BuildingBranch ?? context.HeadCanonicalName;

This would mean that the tests would currently fail when executed locally.

* Fix xml doc

* Add some minor improvements

* Fix packed-refs handling

* Make LibGit2Context behave like ManagedGit

Return null if no HEAD could be determined, return an empty collection if there are no matching tags.

* Add [Ignore] to exclude BuildingTags from CloudBuildAllVars

* Cache tags collections

* Reduce allocations for `GitAnnotatedTag` equality checks

* Skip tag candidate collection, reduce string operations

* Fix handling of peel lines in packed-refs

Without this fix we might have considered one level of transitive annotated tags, if a peel line is present in packed-refs. With this fix we also avoid reading the git pack of annotated tags that have peel lines and do not match the object id we are looking for.

* Fix EnumeratePackedRefsWithPeelLines

* Clarify EnumeratePackedRefsWithPeelLines

* Speedup: Use packed-refs header, consider records peeled if applicable

x

* Fix: read from disposed StreamReader

* Avoid exploring tags unless version.json needs it

* Add first Tags test

* Add annotated tag test

* Capture failure drop in its entirety

* Add runtime check on ParseHex argument

* Collect .git directory too

* Adapt to changes on main

* Touch-ups

---------

Co-authored-by: Andrew Arnott <andrew.arnott@microsoft.com>
Co-authored-by: Andrew Arnott <andrewarnott@live.com>